+++ messages/ru-RU/app/modules/media.php
@@ -2,4 +2,98 @@
 
 return [
 
+    'Media' => "Медиа",
+    'Media library' => "Медиа библиотека",
+
+    'ID' => "ИД",
+    'Name' => "Название",
+    'Alias' => "Алиас",
+    'Image' => "Изображение",
+    'Title' => "Заголовок",
+    'Description' => "Описание",
+    'Keywords' => "Ключевые слова",
+    'Status' => "Статус",
+    'Route' => "Роутинг",
+    'Created' => "Создано",
+    'Updated' => "Обновлено",
+    'Created by' => "Автор создания",
+    'Updated by' => "Автор изменений",
+    'Actions' => "Действия",
+    'All statuses' => "Все статусы",
+    'Draft' => "Черновик",
+    'Published' => "Опубликовано",
+    'Add new post' => "Добавить публикацию",
+    '&larr; Back to list' => "&larr; Назад к списку",
+    'Save' => "Сохранить",
+    'Source' => "Исходник",
+    'Media URL' => "URL-адрес медиа",
+    'Edit' => "Редактирование",
+    'Update' => "Редактировать",
+    'All media' => "Все медиа",
+    'Updating media: {name}' => "Редактирование медиа: {name}",
+    'View media' => "Просмотр медиа",
+
+    'Preview' => "Превью",
+    'Categories' => "Категории",
+    'Categories of media' => "Категории медиа",
+
+    '-- All categories --' => "-- Все категории --",
+    '-- Root category --' => "-- Корневая категория --",
+
+    'All categories' => "Все категории",
+    'New category' => "Новая категория",
+    'Parent category' => "Родительская категория",
+    'View category' => "Просмотр категории",
+    'Add new category' => "Добавить категорию",
+
+    'Updating category: {name}' => "Редактирование категории: {name}",
+    'Category URL' => "URL категории",
+    'Category' => "Категория",
+    'Categories' => "Категории",
+
+    'Caption' => "Подпись",
+    'Alternate' => "Альтернативное название",
+
+    'Add new media' => "Добавить медиа",
+    'Uncategorized' => "Без категории",
+    'Items' => "Элементы",
+    'Media items' => "Медиа элементы",
+
+    'Upload' => "Загрузить",
+    'Upload new media' => "Загрузка новых медиа",
+    'Close' => "Закрыть",
+
+    'Files' => "Файлы",
+    'File size' => "Размер фала",
+    'Mime type' => "MIME тип",
+    'Uploaded by' => "Загружено",
+
+    'The value of field `{attribute}` must be a valid JSON, error: {error}.' => "Значение атрибута `{attribute}` должно быть валидным JSON, ошибка: {error}.",
+    'Param attribute must be unique.' => "Параметр атрибута должен быть уникальным.",
+    'It allowed only Latin alphabet, numbers and the «-», «_» characters.' => "Разрешена только латинница, цифры и «-», «_» символы.",
+
+    'Category has been successfully added!' => "Категория была успешно добавлена!",
+    'An error occurred while add the new category.' => "Произошла ошбка при добавлении новой категории.",
+    'OK! Category `{name}` successfully updated.' => "OK! Категория была успешно обновлена.",
+    'An error occurred while update a category `{name}`.' => "Произошла ошбка при обновлении категории `{name}`.",
+    'OK! Category `{name}` successfully deleted.' => "OK! Категория была успешно удалена.",
+    'An error occurred while deleting a category `{name}`.' => "Произошла ошбка при удалении категории `{name}`.",
+    'The requested category does not exist.' => "Запрошенная категория не существует.",
+
+    'Media item has been successfully added!' => "Медиа элемент был успешно добавлен!",
+    'An error occurred while add the new post.' => "Произошла ошибка при добавлении медиа элемента публикации.",
+    'OK! Blog item `{name}` successfully updated.' => "OK! Медиа элемент был успешно обновлён.",
+    'An error occurred while update a media item `{name}`.' => "Произошла ошибка при обновлении медиа элемента `{name}`.",
+    'OK! Blog item `{name}` successfully deleted.' => "OK! Медиа элемент был успешно удалён.",
+    'An error occurred while deleting a media item `{name}`.' => "Произошла ошибка при удалении медиа элемента `{name}`.",
+    'The requested media item does not exist.' => "Запрошенный медиа элемент не существует.",
+
+    'default' => "по-умолчанию",
+
+    'Media item has been successfully added!' => "Публикация успешно добавлена!",
+    'OK! Blog item `Тестовая публикация` successfully updated.' => "Публикация успешно добавлена!",
+
+    'No media items available' => "Нет доступных для отображения медиа элементов",
+    'An error occurred while retrieving the posts list' => "Произошла ошибка при получении списка публикаций",
+
 ];