+++ views/cats/create.php
@@ -3,7 +3,7 @@
 use yii\helpers\Html;
 
 /* @var $this yii\web\View */
-/* @var $model wdmg\blog\models\Categories */
+/* @var $model wdmg\media\models\Categories */
 
 $this->title = Yii::t('app/modules/media', 'New category');
 $this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/media', 'Media library'), 'url' => ['list/index']];
@@ -13,7 +13,7 @@ $this->params['breadcrumbs'][] = $this->title;
 <div class="page-header">
     <h1><?= Html::encode($this->title) ?><small class="text-muted pull-right">[v.<?= $this->context->module->version ?>]</small></h1>
 </div>
-<div class="blog-cats-create">
+<div class="media-cats-create">
     <?= $this->render('_form', [
         'module' => $module,
         'model' => $model,